refactor(chat-core): migrate getSessionBootstrap / getSessionHistory / getMessageRuntimeDetails into ChatRuntime
