+++ packages/chat-core/src/runtime.ts
@@ -119,6 +119,14 @@ import type {
 import type { ChatSessionRow, SessionStore } from './session-port.js';
 import type { StreamBuffer, StreamEventTypeName } from './stream-port.js';
 import type { MeshDispatchPort } from './mesh-port.js';
+import {
+  buildAssistantMessageHistoryDetails,
+  buildChatHistoryTimeline,
+  compactChatHistoryTimelineForSummary,
+  type ChatHistoryMessage,
+  type ChatHistoryStreamEvent,
+  type ChatHistoryTimelineItem,
+} from './history.js';
 
 /**
  * Local tool definition input — structural shape mirroring
@@ -357,6 +365,56 @@ export type ChatBootstrapStreamEvent = {
   data: unknown;
   sequence: number;
   createdAt: Date;
+};
+
+/**
+ * Lot 14b — options for `ChatRuntime.getSessionBootstrap`. Mirrors the
+ * existing `ChatService.getSessionBootstrap` shape verbatim.
+ */
+export type GetSessionBootstrapOptions = {
+  readonly sessionId: string;
+  readonly userId: string;
+};
+
+export type GetSessionBootstrapResult = {
+  readonly messages: ChatMessageWithFeedback[];
+  readonly todoRuntime: Record<string, unknown> | null;
+  readonly checkpoints: ChatCheckpointSummary[];
+  readonly documents: ReadonlyArray<ChatSessionDocumentItem>;
+  readonly assistantDetailsByMessageId: Record<string, ChatBootstrapStreamEvent[]>;
+};
+
+/**
+ * Lot 14b — options for `ChatRuntime.getSessionHistory`. Mirrors the
+ * existing `ChatService.getSessionHistory` shape verbatim.
+ */
+export type GetSessionHistoryOptions = {
+  readonly sessionId: string;
+  readonly userId: string;
+  readonly detailMode?: 'summary' | 'full';
+};
+
+export type GetSessionHistoryResult = {
+  readonly sessionId: string;
+  readonly title: string | null;
+  readonly todoRuntime: Record<string, unknown> | null;
+  readonly checkpoints: ChatCheckpointSummary[];
+  readonly documents: ReadonlyArray<ChatSessionDocumentItem>;
+  readonly items: ChatHistoryTimelineItem[];
+};
+
+/**
+ * Lot 14b — options for `ChatRuntime.getMessageRuntimeDetails`. Mirrors
+ * the existing `ChatService.getMessageRuntimeDetails` shape verbatim.
+ */
+export type GetMessageRuntimeDetailsOptions = {
+  readonly messageId: string;
+  readonly userId: string;
+};
+
+export type GetMessageRuntimeDetailsResult = {
+  readonly messageId: string;
+  readonly items: ChatHistoryTimelineItem[];
 };
 
 /**
@@ -1460,6 +1518,197 @@ export class ChatRuntime {
     return {
       messages,
       todoRuntime,
+    };
+  }
+
+  /**
+   * BR14b Lot 14b — verbatim port of `ChatService.getSessionBootstrap`.
+   * Composes `SessionStore.findForUser` (authz precheck) +
+   * `this.listMessages` (Lot 14a) + `this.listCheckpoints` (Lot 11) +
+   * `deps.resolveSessionWorkspaceId` + `deps.listSessionDocuments` +
+   * `deps.listAssistantDetailsByMessageId` callbacks (Lot 14b).
+   *
+   * Differences from the pre-Lot 14b chat-service body are limited to:
+   *   (a) `this.getSessionForUser` → `this.deps.sessionStore.findForUser`;
+   *   (b) `this.listMessages` → `this.listMessages` (runtime own method);
+   *   (c) `this.listCheckpoints({sessionId, userId, limit})` →
+   *       `this.listCheckpoints({sessionId, limit})` because the runtime
+   *       method does not re-do the authz check; the composer guards
+   *       upfront via `findForUser`;
+   *   (d) `this.resolveSessionWorkspaceId` →
+   *       `this.deps.resolveSessionWorkspaceId`;
+   *   (e) `this.listSessionDocuments` → `this.deps.listSessionDocuments`;
+   *   (f) `this.listAssistantDetailsByMessageId` →
+   *       `this.deps.listAssistantDetailsByMessageId`.
+   */
+  async getSessionBootstrap(
+    options: GetSessionBootstrapOptions,
+  ): Promise<GetSessionBootstrapResult> {
+    const session = await this.deps.sessionStore.findForUser(
+      options.sessionId,
+      options.userId,
+    );
+    if (!session) throw new Error('Session not found');
+
+    const [{ messages, todoRuntime }, checkpoints, workspaceId] = await Promise.all([
+      this.listMessages(options.sessionId, options.userId),
+      this.listCheckpoints({
+        sessionId: options.sessionId,
+        limit: 20,
+      }),
+      this.deps.resolveSessionWorkspaceId(session, options.userId),
+    ]);
+
+    const documents = await this.deps.listSessionDocuments({
+      sessionId: options.sessionId,
+      workspaceId,
+    });
+    const assistantMessageIds = messages
+      .filter((message) => message.role === 'assistant')
+      .map((message) => String(message.id ?? '').trim())
+      .filter((messageId) => messageId.length > 0);
+    const assistantDetailsByMessageId =
+      await this.deps.listAssistantDetailsByMessageId(assistantMessageIds);
+
+    return {
+      messages,
+      todoRuntime,
+      checkpoints,
+      documents,
+      assistantDetailsByMessageId,
+    };
+  }
+
+  /**
+   * BR14b Lot 14b — verbatim port of `ChatService.getSessionHistory`.
+   * Composes the same persistence pieces as `getSessionBootstrap` plus
+   * the pure history projection helpers
+   * (`buildChatHistoryTimeline` + `compactChatHistoryTimelineForSummary`)
+   * moved into `packages/chat-core/src/history.ts` in the same lot.
+   *
+   * Differences from the pre-Lot 14b chat-service body match
+   * `getSessionBootstrap` plus:
+   *   (g) `buildChatHistoryTimeline` /
+   *       `compactChatHistoryTimelineForSummary` resolve from
+   *       `./history.js` (chat-core) instead of `./chat-session-history`
+   *       (the api re-export shim keeps existing api-side callers
+   *       working).
+   */
+  async getSessionHistory(
+    options: GetSessionHistoryOptions,
+  ): Promise<GetSessionHistoryResult> {
+    const session = await this.deps.sessionStore.findForUser(
+      options.sessionId,
+      options.userId,
+    );
+    if (!session) throw new Error('Session not found');
+
+    const [{ messages, todoRuntime }, checkpoints, workspaceId] = await Promise.all([
+      this.listMessages(options.sessionId, options.userId),
+      this.listCheckpoints({
+        sessionId: options.sessionId,
+        limit: 20,
+      }),
+      this.deps.resolveSessionWorkspaceId(session, options.userId),
+    ]);
+
+    const documents = await this.deps.listSessionDocuments({
+      sessionId: options.sessionId,
+      workspaceId,
+    });
+    const assistantMessageIds = messages
+      .filter((message) => message.role === 'assistant')
+      .map((message) => String(message.id ?? '').trim())
+      .filter((messageId) => messageId.length > 0);
+    const assistantDetailsByMessageId =
+      await this.deps.listAssistantDetailsByMessageId(assistantMessageIds);
+    const eventMap = new Map<string, ChatHistoryStreamEvent[]>();
+    for (const [messageId, events] of Object.entries(assistantDetailsByMessageId)) {
+      eventMap.set(messageId, events as ChatHistoryStreamEvent[]);
+    }
+    const projectedItems = buildChatHistoryTimeline(
+      messages as ChatHistoryMessage[],
+      eventMap,
+    );
+    const items =
+      options.detailMode === 'full'
+        ? projectedItems
+        : compactChatHistoryTimelineForSummary(projectedItems);
+
+    return {
+      sessionId: options.sessionId,
+      title: session.title ?? null,
+      todoRuntime,
+      checkpoints,
+      documents,
+      items: [...items].reverse(),
+    };
+  }
+
+  /**
+   * BR14b Lot 14b — verbatim port of
+   * `ChatService.getMessageRuntimeDetails`. Composes
+   * `MessageStore.findDetailedForUser` (authz + role precheck) +
+   * `this.listMessages` (Lot 14a) +
+   * `deps.listAssistantDetailsByMessageId` callback (Lot 14b) + the
+   * pure projection helpers from `./history.js`.
+   *
+   * Differences from the pre-Lot 14b chat-service body are limited to:
+   *   (a) `this.getDetailedMessageForUser` →
+   *       `this.deps.messageStore.findDetailedForUser`;
+   *   (b) `this.listMessages` → `this.listMessages` (runtime own method);
+   *   (c) `this.listAssistantDetailsByMessageId` →
+   *       `this.deps.listAssistantDetailsByMessageId`;
+   *   (d) `buildChatHistoryTimeline` /
+   *       `buildAssistantMessageHistoryDetails` resolve from
+   *       `./history.js` (chat-core) instead of `./chat-session-history`.
+   */
+  async getMessageRuntimeDetails(
+    options: GetMessageRuntimeDetailsOptions,
+  ): Promise<GetMessageRuntimeDetailsResult> {
+    const message = await this.deps.messageStore.findDetailedForUser(
+      options.messageId,
+      options.userId,
+    );
+    if (!message) throw new Error('Message not found');
+    if (message.role !== 'assistant') {
+      throw new Error('Runtime details only exist for assistant messages');
+    }
+
+    const { messages } = await this.listMessages(
+      message.sessionId,
+      options.userId,
+    );
+    const details = await this.deps.listAssistantDetailsByMessageId([options.messageId]);
+    const events = (details[options.messageId] ?? []) as ChatHistoryStreamEvent[];
+    const eventMap = new Map<string, ChatHistoryStreamEvent[]>();
+    eventMap.set(options.messageId, events);
+    const projected = buildChatHistoryTimeline(
+      messages as ChatHistoryMessage[],
+      eventMap,
+    );
+    const firstIndex = projected.findIndex(
+      (item) => String(item.message.id ?? '').trim() === options.messageId,
+    );
+    const lastIndex = (() => {
+      for (let index = projected.length - 1; index >= 0; index -= 1) {
+        if (String(projected[index]?.message.id ?? '').trim() === options.messageId) {
+          return index;
+        }
+      }
+      return -1;
+    })();
+    const items =
+      firstIndex >= 0 && lastIndex >= firstIndex
+        ? projected.slice(firstIndex, lastIndex + 1)
+        : buildAssistantMessageHistoryDetails(
+            message as ChatHistoryMessage,
+            events,
+          );
+
+    return {
+      messageId: options.messageId,
+      items,
     };
   }
 }
